requirements.txt
Flask==2.2.3
pyTelegramBotAPI==4.11.0
python-dotenv==1.0.0
